call `TableNames.lua` to get table names
