+++ app/process/bicycleParking/bicycleParking.lua
@@ -1,5 +1,4 @@
 package.path = package.path .. ";/app/process/helper/?.lua"
-require("DiffTables")
 require("CopyTags")
 require("Metadata")
 require("Set")
