+++ app/process/bikeroutes/bikeroutes.lua
@@ -1,5 +1,4 @@
 package.path = package.path .. ";/app/process/helper/?.lua;"
-require("DiffTables")
 require("CopyTags")
 require("Set")
 require("Sanitize")
