+++ app/process/places/places_todoList.lua
@@ -1,5 +1,4 @@
 package.path = package.path .. ";/app/process/helper/?.lua"
-require("DiffTables")
 require("Set")
 require("Metadata")
 require("CopyTags")
