+++ app/process/poiClassification/poiClassification.lua
@@ -1,5 +1,4 @@
 package.path = package.path .. ";/app/process/helper/?.lua;/app/process/poiClassification/?.lua"
-require("DiffTables")
 require("Set")
 require("ExtractKeys")
 require("InferAddress")
